refactor: config, config_server => local_config, shared_config
* Clear up the semantics and code locations of gateway.etcd & gateway.config modules.
* This will allow (auto-)reloading of configurations in the future.
