+++ setup.cfg
@@ -42,7 +42,7 @@ install_requires =
     aiopg~=1.0.0
     aioredis~=1.3.1
     aioredlock~=0.5.1
-    aiotools>=0.9.1
+    aiotools>=1.0.0
     alembic~=1.4.2
     async_timeout~=3.0
     asyncache>=0.1.1
