+++ src/ai/backend/gateway/events.py
@@ -25,8 +25,8 @@ from aiohttp import web
 import aiohttp_cors
 from aiohttp_sse import sse_response
 import aioredis
-from aiotools import (
-    adefer,
+from aiotools import adefer
+from aiotools.taskgroup import (
     TaskGroup,
     TaskGroupError,
 )
@@ -41,17 +41,16 @@ from ai.backend.common.types import (
     aobject,
     AgentId,
 )
-from ai.backend.common.utils import current_loop
 from .auth import auth_required
 from .defs import REDIS_STREAM_DB
 from .exceptions import GenericNotFound, GenericForbidden, GroupNotFound
 from .manager import READ_ALLOWED, server_status_required
-from .types import CORSOptions, WebMiddleware
 from .utils import check_api_params
 from ..manager.models import kernels, groups, UserRole
 from ..manager.types import BackgroundTaskEventArgs, Sentinel
 if TYPE_CHECKING:
-    from ..gateway.config import LocalConfig
+    from .types import CORSOptions, WebMiddleware
+    from ..gateway.config import LocalConfig, SharedConfig
 
 log = BraceStyleAdapter(logging.getLogger('ai.backend.gateway.events'))
 
@@ -88,7 +87,6 @@ class EventDispatcher(aobject):
     Subscriber example: enqueuing events to the queues for event streaming API handlers
     '''
 
-    loop: asyncio.AbstractEventLoop
     consumers: MutableMapping[str, Set[EventHandler]]
     subscribers: MutableMapping[str, Set[EventHandler]]
     redis_producer: aioredis.Redis
@@ -98,9 +96,9 @@ class EventDispatcher(aobject):
     subscriber_task: asyncio.Task
     producer_lock: asyncio.Lock
 
-    def __init__(self, config: LocalConfig) -> None:
-        self.loop = current_loop()
-        self.config = config
+    def __init__(self, local_config: LocalConfig, shared_config: SharedConfig) -> None:
+        self.local_config = local_config
+        self.shared_config = shared_config
         self.consumers = defaultdict(set)
         self.subscribers = defaultdict(set)
 
@@ -108,12 +106,12 @@ class EventDispatcher(aobject):
         self.redis_producer = await self._create_redis()
         self.redis_consumer = await self._create_redis()
         self.redis_subscriber = await self._create_redis()
-        self.consumer_task = self.loop.create_task(self._consume())
-        self.subscriber_task = self.loop.create_task(self._subscribe())
+        self.consumer_task = asyncio.create_task(self._consume())
+        self.subscriber_task = asyncio.create_task(self._subscribe())
         self.producer_lock = asyncio.Lock()
 
     async def _create_redis(self):
-        redis_url = self.config.get_redis_url(db=REDIS_STREAM_DB)
+        redis_url = self.shared_config.get_redis_url(db=REDIS_STREAM_DB)
         return await redis.connect_with_retries(str(redis_url), encoding=None)
 
     async def close(self) -> None:
@@ -164,8 +162,9 @@ class EventDispatcher(aobject):
                                  args: Tuple[Any, ...] = tuple()) -> None:
         log_fmt = 'DISPATCH_CONSUMERS(ev:{}, ag:{})'
         log_args = (event_name, agent_id)
-        if self.config['debug']['log-events']:
+        if self.local_config['debug']['log-events']:
             log.debug(log_fmt, *log_args)
+        loop = asyncio.get_running_loop()
         async with TaskGroup(name='event-consume-handlers') as task_group:
             for consumer in self.consumers[event_name]:
                 cb = consumer.callback
@@ -176,7 +175,7 @@ class EventDispatcher(aobject):
                         task_group.create_task(cb(consumer.context, agent_id, event_name, *args))
                     else:
                         cb = functools.partial(cb, consumer.context, agent_id, event_name, *args)
-                        self.loop.call_soon(cb)
+                        loop.call_soon(cb)
                 except asyncio.CancelledError:
                     raise
                 except Exception:
@@ -186,8 +185,9 @@ class EventDispatcher(aobject):
                                    args: Tuple[Any, ...] = tuple()) -> None:
         log_fmt = 'DISPATCH_SUBSCRIBERS(ev:{}, ag:{})'
         log_args = (event_name, agent_id)
-        if self.config['debug']['log-events']:
+        if self.local_config['debug']['log-events']:
             log.debug(log_fmt, *log_args)
+        loop = asyncio.get_running_loop()
         async with TaskGroup(name='event-subscribe-handlers') as task_group:
             for subscriber in self.subscribers[event_name]:
                 cb = subscriber.callback
@@ -198,13 +198,14 @@ class EventDispatcher(aobject):
                         task_group.create_task(cb(subscriber.context, agent_id, event_name, *args))
                     else:
                         cb = functools.partial(cb, subscriber.context, agent_id, event_name, *args)
-                        self.loop.call_soon(cb)
+                        loop.call_soon(cb)
                 except asyncio.CancelledError:
                     raise
                 except Exception:
                     log.exception(log_fmt + ': unexpected-error', *log_args)
 
     async def _consume(self) -> None:
+        loop = asyncio.get_running_loop()
         while True:
             try:
                 key, raw_msg = await redis.execute_with_retries(
@@ -214,7 +215,6 @@ class EventDispatcher(aobject):
                                               msg['agent_id'],
                                               msg['args'])
             except TaskGroupError as e:
-                loop = asyncio.get_running_loop()
                 for sub_error in e.__errors__:
                     loop.call_exception_handler({
                         'exception': sub_error,
@@ -234,11 +234,11 @@ class EventDispatcher(aobject):
                                                 msg['agent_id'],
                                                 msg['args'])
 
+        loop = asyncio.get_running_loop()
         while True:
             try:
                 await redis.execute_with_retries(lambda: _subscribe_impl())
             except TaskGroupError as e:
-                loop = asyncio.get_running_loop()
                 for sub_error in e.__errors__:
                     loop.call_exception_handler({
                         'exception': sub_error,
