+++ src/ai/backend/manager/models/base.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 import asyncio
 import collections
 import enum
@@ -155,12 +156,13 @@ class ResourceSlotColumn(TypeDecorator):
             return value
         return value.to_json() if value is not None else None
 
-    def process_result_value(self, value: Dict[str, str], dialect):
+    def process_result_value(self, raw_value: Dict[str, str], dialect):
         # legacy handling
-        mem = value.get('mem')
+        interim_value: Dict[str, Any] = raw_value
+        mem = raw_value.get('mem')
         if isinstance(mem, str) and not mem.isdigit():
-            value['mem'] = BinarySize.from_str(mem)
-        return ResourceSlot.from_json(value) if value is not None else None
+            interim_value['mem'] = BinarySize.from_str(mem)
+        return ResourceSlot.from_json(interim_value) if raw_value is not None else None
 
     def copy(self):
         return ResourceSlotColumn()
