+++ src/ai/backend/manager/models/group.py
@@ -35,7 +35,7 @@ from .base import (
     batch_result,
 )
 from .user import UserRole
-from ai.backend.gateway.config import RESERVED_DOTFILES
+from ..defs import RESERVED_DOTFILES
 from ai.backend.common import msgpack
 
 log = BraceStyleAdapter(logging.getLogger(__file__))
